Reorganize module hierarchy
Correct some paths
